+++ hive/utils/advantage.py
